Button response
Buttons are now hidden and show when the corresponding key is pressed
(arrow keys). This will be tied into Makey Makey implementation where
people sit down and the menu options show.

BUG: fadeOut on food button is not working properly. Alpha keeps
decreasing due to fadeOutTimer not stopping for some reason when
fadeOutTimer.stop() is called. Not sure what the bug is.